LWIS: Run the rest of the commands in cleanup transaction when error
If there is any error occuring when executing the transaction
commands, it will stop and return error to upper layer.
It means the rest of the commands in the transaction won't be
executed. It might cause more issues when the transaction is
doing cleanup. Skip the error and run the rest of the commands
in cleanup transaction to prevent more issues and leakages.

Bug: 200226703
Test: CTS, GCA
Signed-off-by: Holmes Chou <holmeschou@google.com>
Change-Id: Ia1b07b9768d4b2d8e9a41b9f05fd0fe901186951