Buffer: fix dma_buf use after free
We export dma_buf fd to userspace and use get_dma_buf
to increase refcount of the fd to 2 which handles both
of close(fd) and lwis_buffer_free(). To fix this issue
by race condition, we need to make sure we allocated
dma_buf and get fd from dma_buf_fd(dma_buf, O_CLOEXEC) which
will increase the refcount by 1. It's possible to free this fd
by hacker before we use get_dma_buf() to increase refcount of
the fd to 2.

So we need to move get_dma_buf() before dma_buf_fd() and avoid
we just have refcount = 1 case.

Bug: 209436980
Test: test_poc and no memory leakage
Signed-off-by: Nick Chung <nickchung@google.com>
Change-Id: I26e00df2c0657e790d8a9663556778d177d1d313