LWIS: add LWIS_FENCE_ENABLED for zuma
The LWIS fence feature was only added for devices launching with the
5.15 kernel. Since this feature modifies the UAPI by introducing the
variable trigger_condition, we need to compile it out of the P21 driver.
We can do that by moving all the fence related code to lwis_fence.c and
using the LWIS_FENCE_ENABLED define to disable it for P21. Refer to
ag/18085883 for more platform details.

Tested: verify camera on P21
Bug: 229024993
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I2145b6198c137557022649d40881f91cdc1418b1