LWIS: Unifying device power sequences
Sensor modules (which consist of OIS, actuator, image sensor and EEPROM)
usually share the same power sequence. It would be advantageous for us
to unify the definition of the power sequences into one place, and
properly reference count them. So that we won't wasting cycles
performing power up (or down) sequence multiple times.

Bug: 254536545
Test: GCA, CTS
Change-Id: I35b83bd8ee76219127b7b09f4cf2fde9f73e50aa
Signed-off-by: Holmes Chou <holmeschou@google.com>