LWIS: Update read buffer pointer
While executing the remaining transaction based
on the transaction process limit, there is a need
to save the location where the read buffer should
be updated in the response.
Queue the work on the tranaction/I2C worker thread
again if there are any pending entries to be
processed further.

Bug: 284491752
Bug: 279088769
Test: GCA Smoke Test, verified with LDAF changes
Change-Id: I1054318a028e487f94a4f667c589cedb3e4b8e71